Per line-based deduplication in example_molecules.smi
In response to a clarification by Ian A Watson, the larger list of
test molecules is simplified on _per line_-basis with the Emacs
editor:
+ alphabetic sort of the lines, where SMILES string and PubChem
  identifier are considered as an immutable ensemble, by C-x h,
  and M-x sort-lines
+ remove of any blank line present, by C-x h, M-x delete-blank-lines
+ remove of duplicate lines, again SMILES string and PubChem
  identifier per line are considered as immutable ensemble, by
  C-x h and M-x delete-duplicate-lines

This shortens the number of entries from previously 30810 to now
24986 entries without change of the SMILES' syntax.  The test set
is known to contain a few entries where one SMILES string is
attributed more than one PubChem identifier.  The test set equally
is known to contain a few entries where one PubChem identifier is
attributed to more than one SMILES string (pure compound and compound
adduct), too.

Both deviations from a bijective correspondence between a molecular
structure and a 'bookkeeping', structure-insensitive label are beyond the
control of this project, and not unique to PubChem (e.g., CAS registry
numbers attributed to crystalline polymorphs), either.